refactor to use openapi2puml instead of swagger2puml
