+++ README.md
@@ -11,7 +11,7 @@ The original developers of Swagger2puml are:
 This project is based on Maven.
 Following are modules we currently have 
 
-- swagger2puml-core
+- openapi2puml-core
 
 Following are the tools which this project internally uses:
 
@@ -20,21 +20,19 @@ Following are the tools which this project internally uses:
 - [Graphviz]
 - [Mustache]
 
-Please [install](https://graphviz.gitlab.io/download/) Graphviz on your machines, ensure you edit PATH varible of your machine and add bin folder of graphviz.
-
-Create a new System Variable called GRAPHVIZ_DOT and point to dot.exe for windows and dot in case of Linux in bin folder of graphviz.
-
 # How does it work
 
-- Swagger2Puml internally reads the swagger definition using [Swagger Parser] which then converts the Swagger Definition to swagger.puml
-- Once the swagger.puml gets generated sucessfully it then calls [Plant UML] to generate swagger.svg
+- Input: Openapi2Puml parses the swagger definition from input using [Swagger Parser] 
+- Transform: The swagger definition is built into an object model 
+- Output: The object model is transformed into a [Plant UML] file using a [Mustache] template. Optionally a .svg 
+image can be generated also. 
 
 
-## swagger2puml-core: 
+## openapi2puml-core: 
 
-This utility takes Swagger Yaml as input and as response it generates swagger.puml and swagger.svg files as output.
+This utility takes OpenAPI Yaml or JSON as input generates swagger.puml and swagger.svg files as output.
 
-Below is the Sample Class Diagram which gets generated.
+Below is the Sample Class Diagram generated by the application.
 To see the generated PUML file, please click [here](examples/swagger.puml)
 
 ![Swagger-Class-Diagram-Sample](examples/swagger.svg)
@@ -42,13 +40,14 @@ To see the generated PUML file, please click [here](examples/swagger.puml)
 ### Usage:
 
 ```
-java -cp swagger2puml.jar com.kicksolutions.swagger.Swagger2PlantUML [options]
+java -cp openapi2puml.jar org.openapi2puml.openapi.OpenApi2PlantUML [options]
 
 -i {Path of Swagger Definition (Can be either Yaml or json)}
 -o {Target location where Puml File and Image should generated}
 -generateDefinitionModelOnly {true/flase Defult False (Optional)}
 -includeCardinality {true/flase Defult true (Optional)}
--includeCardinality {true/flase Defult true (Optional)}
+-generateSvg true/false; Default=true 
+
 ```
 
 License
