+++ openapi2puml-core/src/test/resources/petstore/swagger.yaml
