+++ openapi2puml-core/src/test/resources/petstore_with_inheritance/swagger.yaml
